assistant/ai_service.py
import google.generativeai as genai
import os
import json
from dotenv import load_dotenv

load_dotenv()

genai.configure(api_key=os.getenv('GEMINI_API_KEY'))
model = genai.GenerativeModel('gemini-2.5-flash')

SYSTEM_PROMPT = """
You are an AI assistant for Vunoh Global, a platform that helps Kenyans in the diaspora manage tasks back home.

When a customer sends a request, you must return a single valid JSON object with exactly these fields:

{
  "intent": one of: send_money, hire_service, verify_document, airport_transfer, check_status,
  "entities": {
    "amount": number or null,
    "currency": string or null,
    "recipient": string or null,
    "location": string or null,
    "service_type": string or null,
    "document_type": string or null,
    "urgency": "high", "medium", or "low",
    "date": string or null,
    "additional_notes": string or null
  },
  "steps": [list of 4 to 6 clear action strings to fulfil this task],
  "messages": {
    "whatsapp": "conversational message with 1-2 emojis, use line breaks, include task code placeholder {task_code}",
    "email": "formal structured email with subject line, include task code placeholder {task_code}",
    "sms": "under 160 characters, include task code placeholder {task_code}"
  }
}

Rules:
- Return ONLY the JSON object. No explanation, no markdown, no code fences.
- The steps must be specific to the intent, not generic.
- For send_money: steps include identity check, recipient confirmation, transfer initiation, confirmation.
- For hire_service: steps include provider matching, scheduling, customer approval, completion sign-off.
- For verify_document: steps include document intake, legal team review, authority check, result report.
- For airport_transfer: steps include flight details confirmation, driver assignment, pickup coordination, completion.
- urgency must always be set based on words like: urgent, asap, immediately, today (high), soon, this week (medium), or default (low).
- Always include {task_code} in all three messages so it can be replaced later.
"""

def process_request(user_message):
    prompt = f"{SYSTEM_PROMPT}\n\nCustomer request: {user_message}"
    
    response = model.generate_content(prompt)
    raw = response.text.strip()

    # Strip markdown fences if Gemini adds them anyway
    if raw.startswith("```"):
        raw = raw.split("```")[1]
        if raw.startswith("json"):
            raw = raw[4:]
    raw = raw.strip()

    data = json.loads(raw)
    return data